add a chore with store

+++ src/api/routes.py
@@ -44,14 +44,16 @@ def createNewUser():
 
 # get request from chore list
 @api.route('/chore', methods=['GET'])
-def getChoresByEmail(): 
+def getChoresByUserEmail(): 
     email = request.args.get("email", None)
     user = User.get_by_email(email)
-    chores = Chore.get_chores_by_user_id(user.id)
-    serialized_chores = []
-    for chore in chores:
-        serialized_chores.append(chore.serialize())
-    return jsonify({"chores" : serialized_chores})
+    if user: 
+        chores = Chore.get_chores_by_user_id(user.id)
+        serialized_chores = []
+        for chore in chores:
+            serialized_chores.append(chore.serialize())
+        return jsonify({"chores" : serialized_chores})
+    return jsonify({"msg": "no user"}), 404
   
  
 
@@ -82,27 +84,25 @@ def postChore():
 
 # delete a chore 
 @api.route('/chore', methods=['DELETE'])
-def delete_chore():
-    email = request.json.get("email", None)
-    chore = request.json.get("chore", None)
-    duration = request.json.get("duration", None)
-    date = request.json.get("date", None)
-
-    user = User.get_by_email(email)
+def deleteChoresByUserEmail():
+    email = request.args.get("email", None)
+    chore_id = request.args.get("chore_id", None)
+    
+    user = User.get_by_email(email.replace("%40", "@"))
     if user:
         try:
-            choreToDelete = Chore.query.filter_by(id="4").first()
+            choreToDelete = Chore.query.filter_by(id=chore_id).first()
         except exc.SQLAlchemyError: 
-            return jsonify("error finding the chore to delete"), 400
-        try:
-            db.session.delete(choreToDelete)
-        except exc.SQLAlchemyError: 
-            return jsonify("error deleting the chore"), 400
-        db.session.commit()
+            return jsonify("error finding the chore to delete"), 404
 
-        return jsonify({"chore": chore, "duration": duration, "date": date, "email": email}), 201
+        if choreToDelete:
+            db.session.delete(choreToDelete)
+            db.session.commit()
+            return jsonify({"msg": "success deleting the chore"}), 201
         
-    return jsonify({"msg": "error adding chore"}), 401
+        return jsonify("There is no chore to delete"), 404
+
+    return jsonify({"msg": "error deleting the chore after 201"}), 401
 
 
 # login end point
